refactor: move slug functions to utils
